pyproject.toml
[tool.poetry]
name = "crypto-news-ingestion-pipeline"
version = "0.1.0"
description = "A Kafka-based data ingestion pipeline for cryptocurrency news."
authors = ["Francisco Rodriguez"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
apscheduler = "^3.10.4"
pydantic-settings = "^2.6.1"
kafka-python = "^2.0.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
